Move Loki package
